refactor: 프로젝트 구조 변경 - 역할에 따라 패키지 구분
