+++ src/main/java/gift/controller/ProductViewController.java
@@ -1,5 +1,6 @@
-package gift;
+package gift.controller;
 
+import gift.DTO.Product;
 import java.util.List;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
